Rename watson mod to watson_extto avoid conflicts
