+++ qwatson/models/tablemodels.py
@@ -21,10 +21,9 @@ from PyQt5.QtCore import (QAbstractTableModel, QModelIndex,
 # ---- Local imports
 
 from qwatson.utils import colors
-from qwatson.utils.dates import (
-    qdatetime_from_str, local_arrow_from_str, contraint_arrow_to_span)
+from qwatson.utils.dates import local_arrow_from_str, contraint_arrow_to_span
 from qwatson.utils.strformating import list_to_str
-from qwatson.watson.watsonhelpers import edit_frame_at
+from qwatson.watson_ext.watsonhelpers import edit_frame_at
 
 
 class WatsonTableModel(QAbstractTableModel):
