+++ qwatson/widgets/tableviews.py
@@ -442,7 +442,7 @@ class FormatedWatsonTableView(BasicWatsonTableView):
 
 
 if __name__ == '__main__':
-    from qwatson.watson.watsonextends import Watson
+    from qwatson.watson_ext.watsonextends import Watson
     from qwatson.models.tablemodels import WatsonTableModel
 
     client = Watson()
